+++ trend-collector/trendcollector/libs/__init__.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env /Users/yasui/Practice/trend_collector/trend-collector/venv/bin/python
 
-from .client import *
+from .services.client import *
 from .twitter_v2 import *
 from .conf import *
 from .response import *
