+++ trend-collector/trendcollector/libs/services/__init__.py
@@ -1,3 +1,4 @@
 from .accessor import *
+from .client import Twitter
 from .collector import TwitterCollector
 from .custom_exception import CustomException
