+++ trend-collector/trendcollector/libs/services/collector.py
@@ -1,7 +1,7 @@
 import abc
 
 from .accessor import TwitterAccountAccessor, TrendAccessor
-from ..client import Twitter
+from ..services.client import Twitter
 from ..models import Trend, TwitterAccount, WoeidRawTrend
